--- core/src/code.rs
@@ -1,988 +1,0 @@
-// This file is part of Gear.
-
-// Copyright (C) 2021-2024 Gear Technologies Inc.
-// SPDX-License-Identifier: GPL-3.0-or-later WITH Classpath-exception-2.0
-
-// This program is free software: you can redistribute it and/or modify
-// it under the terms of the GNU General Public License as published by
-// the Free Software Foundation, either version 3 of the License, or
-// (at your option) any later version.
-
-// This program is distributed in the hope that it will be useful,
-// but WITHOUT ANY WARRANTY; without even the implied warranty of
-// MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE. See the
-// GNU General Public License for more details.
-
-// You should have received a copy of the GNU General Public License
-// along with this program. If not, see <https://www.gnu.org/licenses/>.
-
-//! Module for checked code.
-
-use crate::{
-    ids::CodeId,
-    message::{DispatchKind, WasmEntryPoint},
-    pages::{PageNumber, PageU32Size, WasmPage},
-};
-use alloc::{collections::BTreeSet, vec, vec::Vec};
-use gear_wasm_instrument::{
-    parity_wasm::{
-        self,
-        elements::{
-            ExportEntry, External, GlobalEntry, GlobalType, ImportCountType, InitExpr, Instruction,
-            Internal, Module, Type, ValueType,
-        },
-    },
-    wasm_instrument::gas_metering::{ConstantCostRules, Rules},
-    InstrumentationBuilder, InstrumentationError, SyscallName, STACK_END_EXPORT_NAME,
-};
-use scale_info::{
-    scale::{Decode, Encode},
-    TypeInfo,
-};
-
-/// Defines maximal permitted count of memory pages.
-pub const MAX_WASM_PAGE_COUNT: u16 = 512;
-
-/// Name of exports allowed on chain.
-pub const ALLOWED_EXPORTS: [&str; 6] = [
-    "init",
-    "handle",
-    "handle_reply",
-    "handle_signal",
-    "state",
-    "metahash",
-];
-
-/// Name of exports required on chain (only 1 of these is required).
-pub const REQUIRED_EXPORTS: [&str; 2] = ["init", "handle"];
-
-fn get_exports(module: &Module) -> BTreeSet<DispatchKind> {
-    let mut entries = BTreeSet::new();
-
-    for entry in module
-        .export_section()
-        .expect("Exports section has been checked for already")
-        .entries()
-        .iter()
-    {
-        if let Internal::Function(_) = entry.internal() {
-            if let Some(entry) = DispatchKind::try_from_entry(entry.field()) {
-                entries.insert(entry);
-            }
-        }
-    }
-
-    entries
-}
-
-fn check_exports(module: &Module) -> Result<(), CodeError> {
-    let types = module
-        .type_section()
-        .ok_or(SectionError::NotFound(SectionName::Type))?
-        .types();
-
-    let funcs = module
-        .function_section()
-        .ok_or(SectionError::NotFound(SectionName::Function))?
-        .entries();
-
-    let import_count = module.import_count(ImportCountType::Function) as u32;
-
-    let exports = module
-        .export_section()
-        .ok_or(SectionError::NotFound(SectionName::Export))?
-        .entries();
-
-    let mut entry_point_found = false;
-    for (export_index, export) in exports.iter().enumerate() {
-        let &Internal::Function(func_index) = export.internal() else {
-            continue;
-        };
-
-        let index =
-            func_index
-                .checked_sub(import_count)
-                .ok_or(ExportError::ExportReferencesToImport(
-                    export_index as u32,
-                    func_index,
-                ))?;
-
-        // Panic is impossible, unless the Module structure is invalid.
-        let type_id = funcs
-            .get(index as usize)
-            .unwrap_or_else(|| unreachable!("Module structure is invalid"))
-            .type_ref() as usize;
-
-        // Panic is impossible, unless the Module structure is invalid.
-        let Type::Function(func_type) = types
-            .get(type_id)
-            .unwrap_or_else(|| unreachable!("Module structure is invalid"));
-
-        if !(func_type.params().is_empty() && func_type.results().is_empty()) {
-            Err(ExportError::InvalidExportFnSignature(export_index as u32))?;
-        }
-
-        if !ALLOWED_EXPORTS.contains(&export.field()) {
-            Err(ExportError::ExcessExport(export_index as u32))?;
-        }
-
-        if REQUIRED_EXPORTS.contains(&export.field()) {
-            entry_point_found = true;
-        }
-    }
-
-    entry_point_found
-        .then_some(())
-        .ok_or(ExportError::RequiredExportNotFound)
-        .map_err(CodeError::Export)
-}
-
-fn check_imports(module: &Module) -> Result<(), CodeError> {
-    let types = module
-        .type_section()
-        .ok_or(SectionError::NotFound(SectionName::Type))?
-        .types();
-
-    let imports = module
-        .import_section()
-        .ok_or(SectionError::NotFound(SectionName::Import))?
-        .entries();
-
-    let syscalls = SyscallName::instrumentable_map();
-
-    let mut visited_imports = BTreeSet::new();
-    for (import_index, import) in imports.iter().enumerate() {
-        let External::Function(i) = import.external() else {
-            continue;
-        };
-
-        // Panic is impossible, unless the Module structure is invalid.
-        let Type::Function(func_type) = &types
-            .get(*i as usize)
-            .unwrap_or_else(|| unreachable!("Module structure is invalid"));
-
-        let syscall = syscalls
-            .get(import.field())
-            .ok_or(ImportError::UnknownImport(import_index as u32))?;
-
-        if !visited_imports.insert(*syscall) {
-            Err(ImportError::DuplicateImport(import_index as u32))?;
-        }
-
-        let signature = syscall.signature();
-
-        let params = signature
-            .params()
-            .iter()
-            .copied()
-            .map(Into::<ValueType>::into);
-        let results = signature.results().unwrap_or(&[]);
-
-        if !(params.eq(func_type.params().iter().copied()) && results == func_type.results()) {
-            Err(ImportError::InvalidImportFnSignature(import_index as u32))?;
-        }
-    }
-
-    Ok(())
-}
-
-fn get_export_entry_with_index<'a>(
-    module: &'a Module,
-    name: &str,
-) -> Option<(u32, &'a ExportEntry)> {
-    module
-        .export_section()?
-        .entries()
-        .iter()
-        .enumerate()
-        .find_map(|(export_index, export)| {
-            (export.field() == name).then_some((export_index as u32, export))
-        })
-}
-
-fn get_export_entry_mut<'a>(module: &'a mut Module, name: &str) -> Option<&'a mut ExportEntry> {
-    module
-        .export_section_mut()?
-        .entries_mut()
-        .iter_mut()
-        .find(|export| export.field() == name)
-}
-
-fn get_export_global_with_index(module: &Module, name: &str) -> Option<(u32, u32)> {
-    let (export_index, export) = get_export_entry_with_index(module, name)?;
-    match export.internal() {
-        Internal::Global(index) => Some((export_index, *index)),
-        _ => None,
-    }
-}
-
-fn get_export_global_index_mut<'a>(module: &'a mut Module, name: &str) -> Option<&'a mut u32> {
-    match get_export_entry_mut(module, name)?.internal_mut() {
-        Internal::Global(index) => Some(index),
-        _ => None,
-    }
-}
-
-fn get_init_expr_const_i32(init_expr: &InitExpr) -> Option<i32> {
-    match init_expr.code() {
-        [Instruction::I32Const(const_i32), Instruction::End] => Some(*const_i32),
-        _ => None,
-    }
-}
-
-fn get_global_entry(module: &Module, global_index: u32) -> Option<&GlobalEntry> {
-    module
-        .global_section()?
-        .entries()
-        .get(global_index as usize)
-}
-
-fn get_global_init_const_i32(
-    module: &Module,
-    global_index: u32,
-    export_inedx: u32,
-) -> Result<i32, CodeError> {
-    let init_expr = get_global_entry(module, global_index)
-        .ok_or(ExportError::IncorrectGlobalIndex(
-            global_index,
-            export_inedx,
-        ))?
-        .init_expr();
-    get_init_expr_const_i32(init_expr)
-        .ok_or(InitializationError::StackEnd)
-        .map_err(CodeError::Initialization)
-}
-
-fn check_and_canonize_gear_stack_end(module: &mut Module) -> Result<(), CodeError> {
-    let Some((stack_end_export_index, stack_end_global_index)) =
-        get_export_global_with_index(module, STACK_END_EXPORT_NAME)
-    else {
-        return Ok(());
-    };
-    let stack_end_offset =
-        get_global_init_const_i32(module, stack_end_global_index, stack_end_export_index)?;
-
-    // Checks, that each data segment does not overlap with stack.
-    if let Some(data_section) = module.data_section() {
-        for data_segment in data_section.entries() {
-            let offset = data_segment
-                .offset()
-                .as_ref()
-                .and_then(get_init_expr_const_i32)
-                .ok_or(InitializationError::DataSegment)?;
-
-            if offset < stack_end_offset {
-                Err(StackEndError::StackEndOverlaps)?;
-            }
-        }
-    };
-
-    // If [STACK_END_EXPORT_NAME] points to mutable global, then make new const global
-    // with the same init expr and change the export internal to point to the new global.
-    if get_global_entry(module, stack_end_global_index)
-        .ok_or(ExportError::IncorrectGlobalIndex(
-            stack_end_global_index,
-            stack_end_export_index,
-        ))?
-        .global_type()
-        .is_mutable()
-    {
-        // Panic is impossible, because we have checked above, that global section exists.
-        let global_section = module
-            .global_section_mut()
-            .unwrap_or_else(|| unreachable!("Cannot find global section"));
-        let new_global_index = u32::try_from(global_section.entries().len())
-            .map_err(|_| StackEndError::GlobalIndexOverflow)?;
-        global_section.entries_mut().push(GlobalEntry::new(
-            GlobalType::new(parity_wasm::elements::ValueType::I32, false),
-            InitExpr::new(vec![
-                Instruction::I32Const(stack_end_offset),
-                Instruction::End,
-            ]),
-        ));
-
-        // Panic is impossible, because we have checked above,
-        // that stack end export exists and it points to global.
-        get_export_global_index_mut(module, STACK_END_EXPORT_NAME)
-            .map(|global_index| *global_index = new_global_index)
-            .unwrap_or_else(|| unreachable!("Cannot find stack end export"))
-    }
-
-    Ok(())
-}
-
-/// Section name in WASM module.
-#[derive(Debug, PartialEq, Eq, derive_more::Display)]
-pub enum SectionName {
-    /// Type section.
-    #[display(fmt = "Type section")]
-    Type,
-    /// Import section.
-    #[display(fmt = "Import section")]
-    Import,
-    /// Function section.
-    #[display(fmt = "Function section")]
-    Function,
-    /// Export section.
-    #[display(fmt = "Export section")]
-    Export,
-    /// Start section.
-    #[display(fmt = "Start section")]
-    Start,
-}
-
-/// Section error in WASM module.
-#[derive(Debug, PartialEq, Eq, derive_more::Display)]
-pub enum SectionError {
-    /// Section not found.
-    #[display(fmt = "{_0} not found")]
-    NotFound(SectionName),
-    /// Section not supported.
-    #[display(fmt = "{_0} not supported")]
-    NotSupported(SectionName),
-}
-
-/// Memory error in WASM module.
-#[derive(Debug, PartialEq, Eq, derive_more::Display)]
-pub enum MemoryError {
-    /// Memory entry not found in import section.
-    #[display(fmt = "Memory entry not found")]
-    EntryNotFound,
-    /// The WASM module has invalid count of static memory pages.
-    #[display(fmt = "The WASM module has invalid count of static memory pages")]
-    InvalidStaticPageCount,
-}
-
-/// Stack end error in WASM module.
-#[derive(Debug, PartialEq, Eq, derive_more::Display)]
-pub enum StackEndError {
-    /// Can't insert new global due to index overflow in global section.
-    #[display(fmt = "Can't insert new global due to index overflow")]
-    GlobalIndexOverflow,
-    /// Pointer to the stack end overlaps data segment.
-    #[display(fmt = "Pointer to the stack end overlaps data segment")]
-    StackEndOverlaps,
-}
-
-/// Initialization error in WASM module.
-#[derive(Debug, PartialEq, Eq, derive_more::Display)]
-pub enum InitializationError {
-    /// Unsupported initialization of gear stack end global variable.
-    #[display(fmt = "Unsupported initialization of gear stack end global variable")]
-    StackEnd,
-    /// Unsupported initialization of data segment.
-    #[display(fmt = "Unsupported initialization of data segment")]
-    DataSegment,
-}
-
-/// Export error in WASM module.
-#[derive(Debug, PartialEq, Eq, derive_more::Display)]
-pub enum ExportError {
-    /// Incorrect global export index. Can occur when export refers to not existing global index.
-    #[display(fmt = "Global index `{_0}` in export index `{_1}` is incorrect")]
-    IncorrectGlobalIndex(u32, u32),
-    /// Exporting mutable globals is restricted by the Gear protocol.
-    #[display(fmt = "Global index `{_0}` in export index `{_1}` cannot be mutable")]
-    MutableGlobalExport(u32, u32),
-    /// Export references to an import function, which is not allowed.
-    #[display(fmt = "Export index `{_0}` references to imported function with index `{_1}`")]
-    ExportReferencesToImport(u32, u32),
-    /// The signature of an exported function is invalid.
-    #[display(fmt = "Exported function with index `{_0}` must have signature `fn f() {{ ... }}`")]
-    InvalidExportFnSignature(u32),
-    /// The provided code contains excess function export.
-    #[display(fmt = "Excess export with index `{_0}` found")]
-    ExcessExport(u32),
-    /// The provided code doesn't contain the required `init` or `handle` export function.
-    #[display(fmt = "Required export function `init` or `handle` not found")]
-    RequiredExportNotFound,
-}
-
-/// Import error in WASM module.
-#[derive(Debug, PartialEq, Eq, derive_more::Display)]
-pub enum ImportError {
-    /// The imported function is not supported by the Gear protocol.
-    #[display(fmt = "Unknown imported function with index `{_0}`")]
-    UnknownImport(u32),
-    /// The imported function is declared multiple times.
-    #[display(fmt = "Imported function with index `{_0}` is declared multiple times")]
-    DuplicateImport(u32),
-    /// The signature of an imported function is invalid.
-    #[display(fmt = "Invalid function signature for imported function with index `{_0}`")]
-    InvalidImportFnSignature(u32),
-}
-
-/// Describes why the code is not valid Gear program.
-#[derive(Debug, PartialEq, Eq, derive_more::Display, derive_more::From)]
-pub enum CodeError {
-    /// Validation by wasmparser failed.
-    #[display(fmt = "Wasm validation failed")]
-    Validation,
-    /// Error occurred during decoding original program code.
-    #[display(fmt = "The wasm bytecode is failed to be decoded")]
-    Decode,
-    /// Error occurred during encoding instrumented program.
-    #[display(fmt = "Failed to encode instrumented program")]
-    Encode,
-    /// The provided code contains section error.
-    #[display(fmt = "Section error: {_0}")]
-    Section(SectionError),
-    /// The provided code contains memory error.
-    #[display(fmt = "Memory error: {_0}")]
-    Memory(MemoryError),
-    /// The provided code contains stack end error.
-    #[display(fmt = "Stack end error: {_0}")]
-    StackEnd(StackEndError),
-    /// The provided code contains initialization error.
-    #[display(fmt = "Initialization error: {_0}")]
-    Initialization(InitializationError),
-    /// The provided code contains export error.
-    #[display(fmt = "Export error: {_0}")]
-    Export(ExportError),
-    /// The provided code contains import error.
-    #[display(fmt = "Import error: {_0}")]
-    Import(ImportError),
-    /// Error occurred during instrumentation WASM module.
-    #[display(fmt = "Instrumentation error: {_0}")]
-    Instrumentation(InstrumentationError),
-}
-
-/// Contains instrumented binary code of a program and initial memory size from memory import.
-#[derive(Clone, Debug, PartialEq, Eq)]
-pub struct Code {
-    /// Code instrumented with the latest schedule.
-    code: Vec<u8>,
-    /// The uninstrumented, original version of the code.
-    original_code: Vec<u8>,
-    /// Exports of the wasm module.
-    exports: BTreeSet<DispatchKind>,
-    /// Static pages count from memory import.
-    static_pages: WasmPage,
-    /// Instruction weights version.
-    instruction_weights_version: u32,
-}
-
-fn check_mut_global_exports(module: &Module) -> Result<(), CodeError> {
-    if let (Some(export_section), Some(global_section)) =
-        (module.export_section(), module.global_section())
-    {
-        let global_exports =
-            export_section
-                .entries()
-                .iter()
-                .enumerate()
-                .filter_map(|(export_index, export)| match export.internal() {
-                    Internal::Global(index) => Some((export_index as u32, *index)),
-                    _ => None,
-                });
-
-        for (export_index, global_index) in global_exports {
-            if global_section
-                .entries()
-                .get(global_index as usize)
-                .ok_or(ExportError::IncorrectGlobalIndex(
-                    global_index,
-                    export_index,
-                ))?
-                .global_type()
-                .is_mutable()
-            {
-                Err(ExportError::MutableGlobalExport(global_index, export_index))?;
-            }
-        }
-    }
-
-    Ok(())
-}
-
-fn check_start_section(module: &Module) -> Result<(), CodeError> {
-    if module.start_section().is_some() {
-        log::debug!("Found start section in program code, which is not allowed");
-        Err(SectionError::NotSupported(SectionName::Start))?
-    } else {
-        Ok(())
-    }
-}
-
-/// Configuration for `Code::try_new_mock_`.
-/// By default all checks enabled.
-pub struct TryNewCodeConfig {
-    /// Instrumentation version
-    pub version: u32,
-    /// Stack height limit
-    pub stack_height: Option<u32>,
-    /// Export `STACK_HEIGHT_EXPORT_NAME` global
-    pub export_stack_height: bool,
-    /// Check exports (wasm contains init or handle exports)
-    pub check_exports: bool,
-    /// Check imports (check that all imports are valid syscalls with correct signature)
-    pub check_imports: bool,
-    /// Check and canonize stack end
-    pub check_and_canonize_stack_end: bool,
-    /// Check mutable global exports
-    pub check_mut_global_exports: bool,
-    /// Check start section (not allowed for programs)
-    pub check_start_section: bool,
-    /// Make wasmparser validation
-    pub make_validation: bool,
-}
-
-impl Default for TryNewCodeConfig {
-    fn default() -> Self {
-        Self {
-            version: 1,
-            stack_height: None,
-            export_stack_height: false,
-            check_exports: true,
-            check_imports: true,
-            check_and_canonize_stack_end: true,
-            check_mut_global_exports: true,
-            check_start_section: true,
-            make_validation: true,
-        }
-    }
-}
-
-impl TryNewCodeConfig {
-    /// New default config without exports checks.
-    pub fn new_no_exports_check() -> Self {
-        Self {
-            check_exports: false,
-            ..Default::default()
-        }
-    }
-}
-
-impl Code {
-    /// Create the code by checking and instrumenting `original_code`.
-    fn try_new_internal<R, GetRulesFn>(
-        original_code: Vec<u8>,
-        get_gas_rules: Option<GetRulesFn>,
-        config: TryNewCodeConfig,
-    ) -> Result<Self, CodeError>
-    where
-        R: Rules,
-        GetRulesFn: FnMut(&Module) -> R,
-    {
-        if config.make_validation {
-            wasmparser::validate(&original_code).map_err(|err| {
-                log::trace!("Wasm validation failed: {err}");
-                CodeError::Validation
-            })?;
-        }
-
-        let mut module: Module =
-            parity_wasm::deserialize_buffer(&original_code).map_err(|err| {
-                log::trace!("The wasm bytecode is failed to be decoded: {err}");
-                CodeError::Decode
-            })?;
-
-        if config.check_and_canonize_stack_end {
-            check_and_canonize_gear_stack_end(&mut module)?;
-        }
-        if config.check_mut_global_exports {
-            check_mut_global_exports(&module)?;
-        }
-        if config.check_start_section {
-            check_start_section(&module)?;
-        }
-
-        // get initial memory size from memory import
-        let static_pages = module
-            .import_section()
-            .ok_or(SectionError::NotFound(SectionName::Import))?
-            .entries()
-            .iter()
-            .find_map(|entry| match entry.external() {
-                parity_wasm::elements::External::Memory(mem_ty) => Some(mem_ty.limits().initial()),
-                _ => None,
-            })
-            .map(WasmPage::new)
-            .ok_or(MemoryError::EntryNotFound)?
-            .map_err(|_| MemoryError::InvalidStaticPageCount)?;
-
-        if static_pages.raw() > MAX_WASM_PAGE_COUNT as u32 {
-            Err(MemoryError::InvalidStaticPageCount)?;
-        }
-
-        if config.check_exports {
-            check_exports(&module)?;
-        }
-
-        if config.check_imports {
-            check_imports(&module)?;
-        }
-
-        let exports = get_exports(&module);
-
-        let mut instrumentation_builder = InstrumentationBuilder::new("env");
-
-        if let Some(stack_limit) = config.stack_height {
-            instrumentation_builder.with_stack_limiter(stack_limit, config.export_stack_height);
-        }
-
-        if let Some(get_gas_rules) = get_gas_rules {
-            instrumentation_builder.with_gas_limiter(get_gas_rules);
-        }
-
-        module = instrumentation_builder.instrument(module).map_err(|err| {
-            log::trace!("Failed to instrument program: {err}");
-            CodeError::Instrumentation(err)
-        })?;
-
-        let code = parity_wasm::elements::serialize(module).map_err(|err| {
-            log::trace!("Failed to encode instrumented program: {err}");
-            CodeError::Encode
-        })?;
-
-        Ok(Self {
-            code,
-            original_code,
-            exports,
-            static_pages,
-            instruction_weights_version: config.version,
-        })
-    }
-
-    /// Create the code by checking and instrumenting `original_code`.
-    /// Main logic of instrumentation can be represented by this example:
-    /// Let's take a code:
-    /// ```wasm
-    /// (module
-    ///    (import "env" "memory" (memory 1))
-    ///    (export "init" (func $init))
-    ///    (func $f1
-    ///       <-- f1 code -->
-    ///    )
-    ///    (func $f2
-    ///       if (i32.eqz (i32.const 0))
-    ///          <-- some code -->
-    ///       else
-    ///          <-- some code -->
-    ///       end
-    ///    )
-    ///    (func $f3
-    ///       <-- f3 code -->
-    ///    )
-    ///    (func $init
-    ///       call $f1
-    ///       call $f2
-    ///       call $f3
-    ///       <-- some code -->
-    ///    )
-    /// )
-    /// ```
-    ///
-    /// After instrumentation code will be like:
-    /// ```wasm
-    /// (module
-    ///   (import "env" "memory" (memory 1))
-    ///   (export "init" (func $init_export))
-    ///   (func $gas_charge
-    ///      <-- gas charge impl --> ;; see utils/wasm-instrument/src/lib.rs
-    ///   )
-    ///   (func $f1
-    ///      i32.const 123
-    ///      call $gas_charge
-    ///      <-- f1 code -->
-    ///   )
-    ///   (func $f2
-    ///      i32.const 123
-    ///      call $gas_charge
-    ///      if (i32.eqz (i32.const 0))
-    ///         i32.const 1
-    ///         call $gas_charge
-    ///         <-- some code -->
-    ///      else
-    ///         i32.const 2
-    ///         call $gas_charge
-    ///         <-- some code -->
-    ///      end
-    ///   )
-    ///   (func $init
-    ///      i32.const 123
-    ///      call $gas_charge
-    ///      ;; stack limit check impl see in wasm_instrument::inject_stack_limiter
-    ///      <-- stack limit check and increase -->
-    ///      call $f1
-    ///      <-- stack limit decrease -->
-    ///      <-- stack limit check and increase -->
-    ///      call $f2
-    ///      <-- stack limit decrease -->
-    ///      <-- some code -->
-    ///   )
-    ///   (func $init_export
-    ///      i32.const 123
-    ///      call $gas_charge
-    ///      <-- stack limit check and increase -->
-    ///      call $init
-    ///      <-- stack limit decrease -->
-    ///   )
-    /// )
-    pub fn try_new<R, GetRulesFn>(
-        original_code: Vec<u8>,
-        version: u32,
-        get_gas_rules: GetRulesFn,
-        stack_height: Option<u32>,
-    ) -> Result<Self, CodeError>
-    where
-        R: Rules,
-        GetRulesFn: FnMut(&Module) -> R,
-    {
-        Self::try_new_internal(
-            original_code,
-            Some(get_gas_rules),
-            TryNewCodeConfig {
-                version,
-                stack_height,
-                ..Default::default()
-            },
-        )
-    }
-
-    /// Create new code for mock goals with const or no instrumentation rules.
-    pub fn try_new_mock_const_or_no_rules(
-        original_code: Vec<u8>,
-        const_rules: bool,
-        config: TryNewCodeConfig,
-    ) -> Result<Self, CodeError> {
-        let get_gas_rules = const_rules.then_some(|_module: &Module| ConstantCostRules::default());
-        Self::try_new_internal(original_code, get_gas_rules, config)
-    }
-
-    /// Create new code for mock goals with custom instrumentation rules.
-    pub fn try_new_mock_with_rules<R, GetRulesFn>(
-        original_code: Vec<u8>,
-        get_gas_rules: GetRulesFn,
-        config: TryNewCodeConfig,
-    ) -> Result<Self, CodeError>
-    where
-        R: Rules,
-        GetRulesFn: FnMut(&Module) -> R,
-    {
-        Self::try_new_internal(original_code, Some(get_gas_rules), config)
-    }
-
-    /// Returns the original code.
-    pub fn original_code(&self) -> &[u8] {
-        &self.original_code
-    }
-
-    /// Returns reference to the instrumented binary code.
-    pub fn code(&self) -> &[u8] {
-        &self.code
-    }
-
-    /// Returns wasm module exports.
-    pub fn exports(&self) -> &BTreeSet<DispatchKind> {
-        &self.exports
-    }
-
-    /// Returns instruction weights version.
-    pub fn instruction_weights_version(&self) -> u32 {
-        self.instruction_weights_version
-    }
-
-    /// Returns initial memory size from memory import.
-    pub fn static_pages(&self) -> WasmPage {
-        self.static_pages
-    }
-
-    /// Consumes this instance and returns the instrumented and raw binary codes.
-    pub fn into_parts(self) -> (InstrumentedCode, Vec<u8>) {
-        let original_code_len = self.original_code.len() as u32;
-        (
-            InstrumentedCode {
-                code: self.code,
-                original_code_len,
-                exports: self.exports,
-                static_pages: self.static_pages,
-                version: self.instruction_weights_version,
-            },
-            self.original_code,
-        )
-    }
-}
-
-/// The newtype contains the Code instance and the corresponding id (hash).
-#[derive(Clone, Debug)]
-pub struct CodeAndId {
-    code: Code,
-    code_id: CodeId,
-}
-
-impl CodeAndId {
-    /// Calculates the id (hash) of the raw binary code and creates new instance.
-    pub fn new(code: Code) -> Self {
-        let code_id = CodeId::generate(code.original_code());
-        Self { code, code_id }
-    }
-
-    /// Creates the instance from the precalculated hash without checks.
-    pub fn from_parts_unchecked(code: Code, code_id: CodeId) -> Self {
-        debug_assert_eq!(code_id, CodeId::generate(code.original_code()));
-        Self { code, code_id }
-    }
-
-    /// Returns corresponding id (hash) for the code.
-    pub fn code_id(&self) -> CodeId {
-        self.code_id
-    }
-
-    /// Returns reference to Code.
-    pub fn code(&self) -> &Code {
-        &self.code
-    }
-
-    /// Decomposes this instance.
-    pub fn into_parts(self) -> (Code, CodeId) {
-        (self.code, self.code_id)
-    }
-}
-
-/// The newtype contains the instrumented code and the corresponding id (hash).
-#[derive(Clone, Debug, Decode, Encode, TypeInfo)]
-pub struct InstrumentedCode {
-    code: Vec<u8>,
-    original_code_len: u32,
-    exports: BTreeSet<DispatchKind>,
-    static_pages: WasmPage,
-    version: u32,
-}
-
-impl InstrumentedCode {
-    /// Returns reference to the instrumented binary code.
-    pub fn code(&self) -> &[u8] {
-        &self.code
-    }
-
-    /// Returns the length of the original binary code.
-    pub fn original_code_len(&self) -> u32 {
-        self.original_code_len
-    }
-
-    /// Returns instruction weights version.
-    pub fn instruction_weights_version(&self) -> u32 {
-        self.version
-    }
-
-    /// Returns wasm module exports.
-    pub fn exports(&self) -> &BTreeSet<DispatchKind> {
-        &self.exports
-    }
-
-    /// Returns initial memory size from memory import.
-    pub fn static_pages(&self) -> WasmPage {
-        self.static_pages
-    }
-
-    /// Consumes the instance and returns the instrumented code.
-    pub fn into_code(self) -> Vec<u8> {
-        self.code
-    }
-}
-
-/// The newtype contains the instrumented code and the corresponding id (hash).
-#[derive(Clone, Debug, Decode, Encode)]
-pub struct InstrumentedCodeAndId {
-    code: InstrumentedCode,
-    code_id: CodeId,
-}
-
-impl InstrumentedCodeAndId {
-    /// Returns reference to the instrumented code.
-    pub fn code(&self) -> &InstrumentedCode {
-        &self.code
-    }
-
-    /// Returns corresponding id (hash) for the code.
-    pub fn code_id(&self) -> CodeId {
-        self.code_id
-    }
-
-    /// Consumes the instance and returns the instrumented code.
-    pub fn into_parts(self) -> (InstrumentedCode, CodeId) {
-        (self.code, self.code_id)
-    }
-}
-
-impl From<CodeAndId> for InstrumentedCodeAndId {
-    fn from(code_and_id: CodeAndId) -> Self {
-        let (code, code_id) = code_and_id.into_parts();
-        let (code, _) = code.into_parts();
-        Self { code, code_id }
-    }
-}
-
-#[cfg(test)]
-mod tests {
-    use crate::code::{Code, CodeError, ExportError};
-    use alloc::vec::Vec;
-    use gear_wasm_instrument::wasm_instrument::gas_metering::ConstantCostRules;
-
-    fn wat2wasm(s: &str) -> Vec<u8> {
-        wabt::Wat2Wasm::new()
-            .validate(true)
-            .convert(s)
-            .unwrap()
-            .as_ref()
-            .to_vec()
-    }
-
-    #[test]
-    fn reject_unknown_exports() {
-        const WAT: &str = r#"
-        (module
-            (import "env" "memory" (memory 1))
-            (export "this_import_is_unknown" (func $test))
-            (func $test)
-        )
-        "#;
-
-        let original_code = wat2wasm(WAT);
-
-        assert_eq!(
-            Code::try_new(original_code, 1, |_| ConstantCostRules::default(), None),
-            Err(CodeError::Export(ExportError::ExcessExport(0)))
-        );
-    }
-
-    #[test]
-    fn required_fn_not_found() {
-        const WAT: &str = r#"
-        (module
-            (import "env" "memory" (memory 1))
-            (export "handle_signal" (func $handle_signal))
-            (func $handle_signal)
-        )
-        "#;
-
-        let original_code = wat2wasm(WAT);
-
-        assert_eq!(
-            Code::try_new(original_code, 1, |_| ConstantCostRules::default(), None),
-            Err(CodeError::Export(ExportError::RequiredExportNotFound))
-        );
-    }
-
-    #[test]
-    fn stack_limit_injection_works() {
-        const WAT: &str = r#"
-        (module
-            (import "env" "memory" (memory 1))
-            (export "init" (func $init))
-            (func $init)
-        )
-        "#;
-
-        let original_code = wat2wasm(WAT);
-
-        let _ = Code::try_new(
-            original_code,
-            1,
-            |_| ConstantCostRules::default(),
-            Some(16 * 1024),
-        )
-        .unwrap();
-    }
-}
